# docs/swagger.yaml
basePath: /
definitions:
  handler.AnalyticsResponse:
    properties:
      clicks_by_day:
        additionalProperties:
          type: integer
        example:
          "2024-01-01": 12
          "2024-01-02": 45
        type: object
      clicks_by_user_agent:
        additionalProperties:
          type: integer
        example:
          Mozilla/5.0: 89
          curl/7.68.0: 12
        type: object
      created_at:
        example: "2026-05-07T10:00:00Z"
        type: string
      original_url:
        example: https://example.com/very/long/url
        type: string
      recent_clicks:
        items:
          $ref: '#/definitions/handler.ClickInfo'
        type: array
      short_code:
        example: abc123
        type: string
      total_clicks:
        example: 1234
        type: integer
    type: object
  handler.ClickInfo:
    properties:
      clicked_at:
        example: "2026-05-07T10:00:00Z"
        type: string
      ip_address:
        example: 192.168.1.100
        type: string
      referer:
        example: https://google.com
        type: string
      user_agent:
        example: Mozilla/5.0 (Windows NT 10.0; Win64; x64)
        type: string
    type: object
  handler.CreateURLRequest:
    properties:
      custom_alias:
        example: mylink
        maxLength: 50
        minLength: 3
        type: string
      expires_at:
        example: "2026-12-31T23:59:59Z"
        type: string
      original_url:
        example: https://example.com/very/long/url
        type: string
    required:
    - original_url
    type: object
  handler.CreateURLResponse:
    properties:
      created_at:
        example: "2026-05-07T10:00:00Z"
        type: string
      custom_alias:
        example: mylink
        type: string
      expires_at:
        example: "2026-12-31T23:59:59Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      original_url:
        example: https://example.com/very/long/url
        type: string
      short_code:
        example: abc123
        type: string
      short_url:
        example: https://short.link/abc123
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      details:
        example: link with short_code abc123 does not exist
        type: string
      error:
        example: link not found
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Link shortening API
  license:
    name: MIT-0
    url: https://github.com/aws/mit-0
  title: Link Shortener API
  version: "1.0"
paths:
  /{short_code}:
    get:
      description: Redirects to the original link using a short code
      parameters:
      - description: Link short code
        in: path
        minLength: 1
        name: short_code
        required: true
        type: string
      responses:
        "301":
          description: Redirect to original URL
          schema:
            type: string
        "400":
          description: Missing short_code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Link not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Follow a short link
      tags:
      - ShortURL
  /{short_code}/analytics:
    get:
      consumes:
      - application/json
      description: Returns click statistics for a short code
      parameters:
      - description: Link short code
        in: path
        minLength: 1
        name: short_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Analytics received
          schema:
            $ref: '#/definitions/handler.AnalyticsResponse'
        "400":
          description: Missing short_code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Link not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get analytics for a short link
      tags:
      - ShortURL
  /health:
    get:
      description: Return service status and current timestamp. No authentication
        required.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - System
  /shorten:
    post:
      consumes:
      - application/json
      description: Creates a shortened link from the original
      parameters:
      - description: Data for creating a short link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Short link created
          schema:
            $ref: '#/definitions/handler.CreateURLResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a shortened link
      tags:
      - ShortURL
swagger: "2.0"